#115: fix rebasing issues
