+++ tests/romtools/test_vector_space.py
@@ -2,7 +2,7 @@ import copy
 import pytest
 import numpy as np
 import romtools as rt
-import romtools.vector_space.utils as utils
+import romtools.trial_space.utils as utils
 
 
 
@@ -23,8 +23,9 @@ def _tensor_to_matrix(tensor_input):
 @pytest.mark.mpi_skip
 def test_dictionary_vector_space():
     snapshots = np.random.normal(size=(3, 8, 6))
+    original_snapshots = snapshots.copy()
     # default test
-    my_shifter = utils.NoOpShifter()
+    my_shifter = utils.create_noop_shifter(snapshots)
     my_splitter = utils.NoOpSplitter()
     my_orthogonalizer = utils.NoOpOrthogonalizer()
     my_vector_space = rt.DictionaryVectorSpace(snapshots,
@@ -37,7 +38,7 @@ def test_dictionary_vector_space():
     assert np.allclose(my_vector_space.get_dimension(), 6)
 
     # test with a shift
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_splitter = utils.NoOpSplitter()
     my_orthogonalizer = utils.NoOpOrthogonalizer()
     my_vector_space = rt.DictionaryVectorSpace(snapshots,
@@ -45,13 +46,13 @@ def test_dictionary_vector_space():
                                              my_splitter,
                                              my_orthogonalizer)
     assert np.allclose(my_vector_space.get_basis().flatten(),
-                       (snapshots - np.mean(snapshots, axis=2)[:, :, None]).flatten())
+                      (original_snapshots - np.mean(original_snapshots, axis=2)[:, :, None]).flatten())
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 6)
 
     # test with a shift and splitting
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     my_orthogonalizer = utils.NoOpOrthogonalizer()
     my_vector_space = rt.DictionaryVectorSpace(snapshots,
@@ -63,7 +64,7 @@ def test_dictionary_vector_space():
     assert np.allclose(my_vector_space.get_dimension(), 12)
 
     # test with a shift, splitting, and orthogonalization
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     my_orthogonalizer = utils.EuclideanL2Orthogonalizer()
     my_vector_space = rt.DictionaryVectorSpace(snapshots,
@@ -74,13 +75,14 @@ def test_dictionary_vector_space():
                        np.mean(snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 12)
     basis = my_vector_space.get_basis()
-    basis = _tensor_to_matrix(basis)
+    basis = tensor_to_matrix(basis)
     assert np.allclose(basis.transpose() @ basis, np.eye(12))
 
 
 @pytest.mark.mpi_skip
 def test_vector_space_from_pod():
     snapshots = np.random.normal(size=(3, 8, 6))
+    original_snapshots = snapshots.copy()
     my_vector_space = rt.VectorSpaceFromPOD(snapshots)
     # truth vector space
     snapshotMatrix = _tensor_to_matrix(snapshots)
@@ -91,28 +93,28 @@ def test_vector_space_from_pod():
     assert np.allclose(0, my_vector_space.get_shift_vector())
 
     # test with a shift
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_vector_space = rt.VectorSpaceFromPOD(snapshots, shifter=my_shifter)
     u, s, v = np.linalg.svd(snapshotMatrix - np.mean(snapshotMatrix, axis=1)[:, None], full_matrices=False)
     basis_tensor = my_vector_space.get_basis()
     assert np.allclose(u.reshape(basis_tensor.shape), basis_tensor)
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 6)
 
     # test with a shift and splitting
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(original_snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     my_vector_space = rt.VectorSpaceFromPOD(snapshots, shifter=my_shifter, splitter=my_splitter)
     u, s, v = np.linalg.svd(my_splitter(snapshotMatrix - np.mean(snapshotMatrix, axis=1)[:, None]), full_matrices=False)
     basis_tensor = my_vector_space.get_basis()
     assert np.allclose(u.reshape(basis_tensor.shape), basis_tensor)
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 12)
 
     # test with a shift, splitting, and orthogonalization
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(original_snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     weighting = np.abs(np.random.normal(size=24))
     my_orthogonalizer = utils.EuclideanVectorWeightedL2Orthogonalizer(weighting)
@@ -122,15 +124,16 @@ def test_vector_space_from_pod():
     basis_tensor = my_vector_space.get_basis()
     assert np.allclose(u.reshape(basis_tensor.shape), basis_tensor)
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 12)
 
 
 @pytest.mark.mpi_skip
-def test_vector_space_from_scaled_pod():
+def test_trial_space_from_scaled_pod():
     snapshots = np.random.normal(size=(3, 8, 6))
+    my_shifter = utils.create_noop_shifter(snapshots)
     my_scaler = utils.VariableScaler('max_abs')
-    my_vector_space = rt.VectorSpaceFromPOD(copy.deepcopy(snapshots), scaler=my_scaler)
+    my_vector_space = rt.VectorSpaceFromPOD(copy.deepcopy(snapshots), shifter=my_shifter, scaler=my_scaler)
     scaled_snapshots = my_scaler.pre_scaling(snapshots)
     snapshotMatrix = _tensor_to_matrix(scaled_snapshots)
     u, s, v = np.linalg.svd(snapshotMatrix, full_matrices=False)
@@ -143,7 +146,8 @@ def test_vector_space_from_scaled_pod():
 
     # test with a shift
     snapshots = np.random.normal(size=(3, 8, 6))
-    my_shifter = utils.AverageShifter()
+    original_snapshots = snapshots.copy()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_scaler = utils.VariableScaler('max_abs')
     my_vector_space = rt.VectorSpaceFromPOD(snapshots, shifter=my_shifter, scaler=my_scaler)
     shifted_snapshots, shift_vector = my_shifter(snapshots)
@@ -154,14 +158,15 @@ def test_vector_space_from_scaled_pod():
     u = u.reshape(basis_tensor.shape)
     u = my_scaler.post_scaling(u)
     assert np.allclose(basis_tensor, u)
-    assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+    assert np.allclose(my_vector_space.get_shift_vector(),+
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 6)
 
     # test with a shift and splitting
     snapshots = np.random.normal(size=(3, 8, 6))
+    original_snapshots = snapshots.copy()
     my_scaler = utils.VariableScaler('max_abs')
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     my_vector_space = rt.VectorSpaceFromPOD(snapshots, shifter=my_shifter, splitter=my_splitter, scaler=my_scaler)
     shifted_snapshots, _ = my_shifter(snapshots)
@@ -174,21 +179,22 @@ def test_vector_space_from_scaled_pod():
     u = my_scaler.post_scaling(u)
     assert np.allclose(basis_tensor, u)
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 12)
 
     # test with a shift, splitting, and orthogonalization
     snapshots = np.random.normal(size=(3, 8, 6))
+    original_snapshots = snapshots.copy()
     my_scaler = utils.VariableScaler('max_abs')
-    my_shifter = utils.AverageShifter()
+    my_shifter = utils.create_average_shifter(snapshots)
     my_splitter = utils.BlockSplitter([[0], [1, 2]], 3)
     weighting = np.abs(np.random.normal(size=24))
     my_orthogonalizer = utils.EuclideanVectorWeightedL2Orthogonalizer(weighting)
     my_vector_space = rt.VectorSpaceFromPOD(snapshots, shifter=my_shifter, splitter=my_splitter, scaler=my_scaler, orthogonalizer=my_orthogonalizer)
     shifted_snapshots, shift_vector = my_shifter(snapshots)
     my_scaler = utils.VariableScaler('max_abs')
-    scaled_shifted_snapshots = my_scaler.pre_scaling(shifted_snapshots)
-    snapshot_matrix = _tensor_to_matrix(scaled_shifted_snapshots)
+    scaled_shifted_snapshots = my_scaler.pre_scaling(snapshots)
+    snapshot_matrix = tensor_to_matrix(scaled_shifted_snapshots)
     snapshot_matrix = my_splitter(snapshot_matrix)
     u, s, v = np.linalg.svd(snapshot_matrix, full_matrices=False)
     ushp = u.shape
@@ -199,7 +205,7 @@ def test_vector_space_from_scaled_pod():
     u = u.reshape(basis_tensor.shape)
     assert np.allclose(basis_tensor, u)
     assert np.allclose(my_vector_space.get_shift_vector(),
-                       np.mean(snapshots, axis=2))
+                       np.mean(original_snapshots, axis=2))
     assert np.allclose(my_vector_space.get_dimension(), 12)
